InfoPage changed to IntroPage. MUI theme implemented. Route updated in app.jsx.
